+++ Sources/Navigator/EPUB/Scripts/src/utils.js
@@ -75,15 +75,16 @@ function appendVirtualColumnIfNeeded() {
   }
 }
 
-var last_known_scrollX_position = 0;
-var last_known_scrollY_position = 0;
+var lastKnownProgressions;
 var ticking = false;
 var viewportWidth = 0;
 
-// Position in range [0 - 1].
-function update(position) {
-  var positionString = position.toString();
-  webkit.messageHandlers.progressionChanged.postMessage(positionString);
+/**
+ * First and last progressions in range [0 - 1].
+ * Expects an object {first, last}
+ */
+function notifyProgressions(progressions) {
+  webkit.messageHandlers.progressionChanged.postMessage(progressions);
 }
 
 window.addEventListener("scroll", onScroll);
@@ -93,28 +94,37 @@ function onScroll() {
     return;
   }
 
-  last_known_scrollY_position =
-    window.scrollY / document.scrollingElement.scrollHeight;
-  // Using Math.abs because for RTL books, the value will be negative.
-  last_known_scrollX_position = Math.abs(
-    window.scrollX / document.scrollingElement.scrollWidth
-  );
+  let root = document.scrollingElement;
+  if (isScrollModeEnabled() && !isVerticalWritingMode()) {
+    const scrollY = window.scrollY;
+    const viewportHeight = window.innerHeight;
+    const totalContentHeight = root.scrollHeight;
+    lastKnownProgressions = {
+      first: scrollY / totalContentHeight,
+      last: (scrollY + viewportHeight) / totalContentHeight,
+    };
+  } else {
+    let scrollX = window.scrollX;
+    const viewportWidth = window.innerWidth;
+    const totalContentWidth = root.scrollWidth;
+
+    if (isRTL()) {
+      scrollX = Math.abs(scrollX);
+    }
+    lastKnownProgressions = {
+      first: scrollX / totalContentWidth,
+      last: (scrollX + viewportWidth) / totalContentWidth,
+    };
+  }
 
   // Window is hidden
-  if (
-    document.scrollingElement.scrollWidth === 0 ||
-    document.scrollingElement.scrollHeight === 0
-  ) {
+  if (root.scrollWidth === 0 || root.scrollHeight === 0) {
     return;
   }
 
   if (!ticking) {
     window.requestAnimationFrame(function () {
-      update(
-        isScrollModeEnabled()
-          ? last_known_scrollY_position
-          : last_known_scrollX_position
-      );
+      notifyProgressions(lastKnownProgressions);
       ticking = false;
     });
   }
@@ -141,6 +151,21 @@ export function isScrollModeEnabled() {
   return style.getPropertyValue("--USER__view").trim() == "readium-scroll-on";
 }
 
+export function isVerticalWritingMode() {
+  const writingMode = window
+    .getComputedStyle(document.documentElement)
+    .getPropertyValue("writing-mode");
+  return writingMode.startsWith("vertical");
+}
+
+export function isRTL() {
+  const style = window.getComputedStyle(document.documentElement);
+  return (
+    style.getPropertyValue("direction") == "rtl" ||
+    style.getPropertyValue("writing-mode") == "vertical-rl"
+  );
+}
+
 // Scroll to the given TagId in document and snap.
 export function scrollToId(id) {
   let element = document.getElementById(id);
@@ -162,8 +187,13 @@ export function scrollToPosition(position, dir) {
   }
 
   if (isScrollModeEnabled()) {
-    let offset = document.scrollingElement.scrollHeight * position;
-    document.scrollingElement.scrollTop = offset;
+    if (!isVerticalWritingMode()) {
+      let offset = document.scrollingElement.scrollHeight * position;
+      document.scrollingElement.scrollTop = offset;
+    } else {
+      let offset = document.scrollingElement.scrollWidth * position;
+      document.scrollingElement.scrollLeft = -offset;
+    }
   } else {
     var documentWidth = document.scrollingElement.scrollWidth;
     var factor = dir == "rtl" ? -1 : 1;
@@ -233,8 +263,8 @@ function scrollToOffset(offset) {
 
 // Snap the offset to the screen width (page width).
 function snapOffset(offset) {
-  var value = offset + 1;
-
+  const delta = isRTL() ? -1 : 1;
+  const value = offset + delta;
   return value - (value % viewportWidth);
 }