+++ Sources/Shared/Toolkit/Format/Sniffers/RARFormatSniffer.swift
@@ -8,6 +8,8 @@ import Foundation
 
 /// Sniffs a RAR file.
 public struct RARFormatSniffer: FormatSniffer {
+    public init() {}
+
     public func sniffHints(_ hints: FormatHints) -> Format? {
         if
             hints.hasFileExtension("rar") ||